Reordenar Entidades de nuestro programa y refactorizar
