+++ src/main/java/modelo/ClienteModel.java
@@ -2,7 +2,7 @@ package modelo;
 
 import DAO.FactoryDAO;
 import DAO.IDao;
-import modelo.cliente.Cliente;
+import modelo.entity.cliente.Cliente;
 
 import java.util.Collection;
 import java.util.HashMap;
