+++ src/main/java/modelo/Modelo.java
@@ -5,7 +5,9 @@ package modelo;
 
 import DAO.FactoryDAO;
 import DAO.IDao;
-import modelo.cliente.Cliente;
+import modelo.entity.Articulo;
+import modelo.entity.Pedido;
+import modelo.entity.cliente.Cliente;
 import modelo.enums.TipoEstado;
 
 import java.util.*;
