+++ src/main/java/modelo/entity/cliente/Cliente.java
@@ -1,6 +1,6 @@
 // Grupo 2 - SQL SQUAD
 
-package modelo.cliente;
+package modelo.entity.cliente;
 
 import jakarta.persistence.*;
 import modelo.enums.TipoCliente;
