+++ src/main/resources/hibernate.cfg.xml
@@ -21,11 +21,11 @@
     <property name="hibernate.connection.autocommit">false</property>
 
     <!-- Mapeo de Entidades -->
-    <mapping class="modelo.Articulo" />
-    <mapping class="modelo.cliente.Cliente" />
-    <mapping class="modelo.cliente.ClienteEstandar" />
-    <mapping class="modelo.cliente.ClientePremium" />
-    <mapping class="modelo.Pedido" />
+    <mapping class="modelo.entity.Articulo" />
+    <mapping class="modelo.entity.cliente.Cliente" />
+    <mapping class="modelo.entity.cliente.ClienteEstandar" />
+    <mapping class="modelo.entity.cliente.ClientePremium" />
+    <mapping class="modelo.entity.Pedido" />
 
   </session-factory>
 </hibernate-configuration>
